Refactor container and fix some batch tests
