+++ shepherd_score/container/_core.py
@@ -2,7 +2,7 @@
 Molecule class to hold molecule geometries and extract interaction profiles.
 MoleculePair class facilitates alignment with interaction profiles.
 """
-from typing import Union, List, Optional, Tuple, Literal
+from typing import Union, List, Optional, Tuple
 from copy import deepcopy
 import sys
 
@@ -381,7 +381,7 @@ class MoleculePair:
                 except ImportError:
                     raise ImportError('jax.numpy and torch is required for this function. Install Jax or just use Torch.')
             import jax.numpy as jnp
-            from .alignment_jax import optimize_ROCS_overlay_jax
+            from shepherd_score.alignment_jax import optimize_ROCS_overlay_jax
             aligned_fit_points, se3_transform, score = optimize_ROCS_overlay_jax(
                 ref_points=jnp.array(ref_atom_pos),
                 fit_points=jnp.array(fit_atom_pos),
@@ -489,7 +489,7 @@ class MoleculePair:
                 except ImportError:
                     raise ImportError('jax.numpy and torch is required for this function. Install Jax or just use Torch.')
             import jax.numpy as jnp
-            from .alignment_jax import optimize_ROCS_esp_overlay_jax
+            from shepherd_score.alignment_jax import optimize_ROCS_esp_overlay_jax
             aligned_fit_points, se3_transform, score = optimize_ROCS_esp_overlay_jax(
                 ref_points=jnp.array(ref_mol_pos),
                 fit_points=jnp.array(fit_mol_pos),
@@ -591,7 +591,7 @@ class MoleculePair:
                 except ImportError:
                     raise ImportError('jax.numpy and torch is required for this function. Install Jax or just use Torch.')
             import jax.numpy as jnp
-            from .alignment_jax import optimize_ROCS_overlay_jax
+            from shepherd_score.alignment_jax import optimize_ROCS_overlay_jax
             aligned_fit_points, se3_transform, score = optimize_ROCS_overlay_jax(
                 ref_points=jnp.array(self.ref_molec.surf_pos),
                 fit_points=jnp.array(self.fit_molec.surf_pos),
@@ -683,7 +683,7 @@ class MoleculePair:
                 except ImportError:
                     raise ImportError('jax.numpy and torch is required for this function. Install Jax or just use Torch.')
             import jax.numpy as jnp
-            from .alignment_jax import optimize_ROCS_esp_overlay_jax
+            from shepherd_score.alignment_jax import optimize_ROCS_esp_overlay_jax
             aligned_fit_points, se3_transform, score = optimize_ROCS_esp_overlay_jax(
                 ref_points=jnp.array(self.ref_molec.surf_pos),
                 fit_points=jnp.array(self.fit_molec.surf_pos),
@@ -784,7 +784,7 @@ class MoleculePair:
                 except ImportError:
                     raise ImportError('jax.numpy and torch is required for this function. Install Jax or just use Torch.')
             import jax.numpy as jnp
-            from .alignment_jax import optimize_esp_combo_score_overlay_jax
+            from shepherd_score.alignment_jax import optimize_esp_combo_score_overlay_jax
             aligned_fit_points, se3_transform, score = optimize_esp_combo_score_overlay_jax(
                 ref_centers_w_H=jnp.array(self.ref_molec.mol.GetConformer().GetPositions()),
                 fit_centers_w_H=jnp.array(self.fit_molec.mol.GetConformer().GetPositions()),
@@ -920,7 +920,7 @@ class MoleculePair:
                 except ImportError:
                     raise ImportError('jax.numpy and torch is required for this function. Install Jax or just use Torch.')
             import jax.numpy as jnp
-            from .alignment_jax import optimize_pharm_overlay_jax, optimize_pharm_overlay_jax_vectorized
+            from shepherd_score.alignment_jax import optimize_pharm_overlay_jax, optimize_pharm_overlay_jax_vectorized
 
             _pharm_fn = optimize_pharm_overlay_jax_vectorized if use_vectorized else optimize_pharm_overlay_jax
             aligned_fit_anchors, aligned_fit_vectors, se3_transform, score = _pharm_fn(
@@ -1282,629 +1282,3 @@ class MoleculePair:
                                          )
         return transformed_fit_molec
 
-
-class MoleculePairBatch:
-    """Batch of MoleculePair objects for fast sequential JAX alignment.
-
-    Pads all atom coordinate arrays to common max shapes so JAX's XLA compiler
-    reuses the same compiled function for every pair, avoiding recompilation.
-    This modifies each MoleculePair in-place (stores results on the pair).
-
-    This is currently optimized for CPU. A GPU-optimized version would
-    benefit from optimizing batches of pairs and using a GPU-optimized alignment.
-    """
-
-    def __init__(self, pairs: List[MoleculePair]):
-        self.pairs = pairs
-
-    def _pad_and_mask_vol(self, no_H: bool = True):
-        """Extract, pad, and create masks for volumetric alignment.
-
-        Does NOT modify the pair objects. Returns padded arrays and masks.
-
-        Parameters
-        ----------
-        no_H : bool
-            If True, use heavy-atom positions (atom_pos). If False, use all atoms.
-
-        Returns
-        -------
-        entries : list of tuples
-            Each tuple is (ref_padded, fit_padded, mask_ref, mask_fit,
-                           orig_ref_len, orig_fit_len).
-        max_ref_len : int
-        max_fit_len : int
-        """
-        if no_H:
-            ref_arrays = [p.ref_molec.atom_pos for p in self.pairs]
-            fit_arrays = [p.fit_molec.atom_pos for p in self.pairs]
-        else:
-            ref_arrays = [p.ref_molec.mol.GetConformer().GetPositions().astype(np.float32)
-                          for p in self.pairs]
-            fit_arrays = [p.fit_molec.mol.GetConformer().GetPositions().astype(np.float32)
-                          for p in self.pairs]
-
-        max_ref_len = max(a.shape[0] for a in ref_arrays)
-        max_fit_len = max(a.shape[0] for a in fit_arrays)
-
-        entries = []
-        for ref_pos, fit_pos in zip(ref_arrays, fit_arrays):
-            orig_ref = ref_pos.shape[0]
-            orig_fit = fit_pos.shape[0]
-
-            ref_padded = np.zeros((max_ref_len, 3), dtype=np.float32)
-            ref_padded[:orig_ref] = ref_pos
-            mask_ref = np.zeros(max_ref_len, dtype=np.float32)
-            mask_ref[:orig_ref] = 1.0
-
-            fit_padded = np.zeros((max_fit_len, 3), dtype=np.float32)
-            fit_padded[:orig_fit] = fit_pos
-            mask_fit = np.zeros(max_fit_len, dtype=np.float32)
-            mask_fit[:orig_fit] = 1.0
-
-            entries.append((ref_padded, fit_padded, mask_ref, mask_fit, orig_ref, orig_fit))
-
-        return entries, max_ref_len, max_fit_len
-
-    def align_with_vol(self,
-                       no_H: bool = True,
-                       num_repeats: int = 50,
-                       trans_init: bool = False,
-                       lr: float = 0.1,
-                       max_num_steps: int = 200,
-                       verbose: bool = False) -> Tuple[np.ndarray, List[np.ndarray]]:
-        """Align all pairs using padded masked volumetric similarity via JAX.
-
-        Because all padded arrays have the same shape, JAX's XLA compiler
-        reuses one compiled kernel for every pair — no recompilation overhead.
-
-        Results are stored in-place on each MoleculePair:
-        - ``pair.transform_vol_noH`` / ``pair.sim_aligned_vol_noH`` (when ``no_H=True``)
-        - ``pair.transform_vol``     / ``pair.sim_aligned_vol``      (when ``no_H=False``)
-
-        Parameters
-        ----------
-        no_H : bool
-            Whether to exclude hydrogens. Default is True.
-        num_repeats : int
-            Number of SE(3) initializations per pair. Default is 50.
-        trans_init : bool
-            If True, initialize translations to each ref atom position. Default is False.
-        lr : float
-            Optimizer learning rate. Default is 0.1.
-        max_num_steps : int
-            Maximum optimization steps. Default is 200.
-        verbose : bool
-            Print scores per pair. Default is False.
-
-        Returns
-        -------
-        scores : np.ndarray
-            Scores for each pair. Shape: (N,).
-        aligned_list : list of np.ndarray
-            Aligned fit atom coordinates (unpadded) for each pair.
-        """
-        try:
-            import jax.numpy as jnp
-        except ImportError as exc:
-            raise ImportError(
-                'JAX is required for MoleculePairBatch.align_with_vol. '
-                'Install it with: pip install "shepherd-score[jax]"'
-            ) from exc
-
-        from .alignment_jax import optimize_ROCS_overlay_jax_mask
-
-        entries, _, _ = self._pad_and_mask_vol(no_H=no_H)
-        aligned_list = []
-        scores = np.zeros((len(self.pairs),))
-
-        for i, (pair, entry) in enumerate(zip(self.pairs, entries)):
-            ref_padded, fit_padded, mask_ref, mask_fit, orig_ref, orig_fit = entry
-
-            trans_centers = None
-            if trans_init:
-                if no_H:
-                    trans_centers = pair.ref_molec.atom_pos
-                else:
-                    trans_centers = pair.ref_molec.mol.GetConformer().GetPositions().astype(np.float32)
-
-            aligned_pts, se3_transform, score = optimize_ROCS_overlay_jax_mask(
-                ref_points=jnp.array(ref_padded),
-                fit_points=jnp.array(fit_padded),
-                mask_ref=jnp.array(mask_ref),
-                mask_fit=jnp.array(mask_fit),
-                alpha=0.81,
-                num_repeats=num_repeats,
-                trans_centers=trans_centers,
-                lr=lr,
-                max_num_steps=max_num_steps,
-                verbose=verbose,
-            )
-
-            se3_transform = np.array(se3_transform)
-            score = np.array(score)
-            aligned_pts = np.array(aligned_pts)[:orig_fit]
-            scores[i] = score
-
-            if no_H:
-                pair.transform_vol_noH = se3_transform
-                pair.sim_aligned_vol_noH = score
-            else:
-                pair.transform_vol = se3_transform
-                pair.sim_aligned_vol = score
-
-            aligned_list.append(aligned_pts)
-
-        return scores, aligned_list
-
-    def _pad_and_mask_vol_esp(self, no_H: bool = True):
-        """Extract, pad, and create masks for volumetric ESP alignment.
-
-        Does NOT modify the pair objects. Returns padded arrays and masks.
-
-        Parameters
-        ----------
-        no_H : bool
-            If True, use heavy-atom positions and non-H partial charges.
-            If False, use all-atom positions and all partial charges.
-
-        Returns
-        -------
-        entries : list of tuples
-            Each tuple is (ref_pos_pad, fit_pos_pad, ref_ch_pad, fit_ch_pad,
-                           mask_ref, mask_fit, orig_ref_len, orig_fit_len).
-        max_ref_len : int
-        max_fit_len : int
-        """
-        if no_H:
-            ref_pos_arrays = [p.ref_molec.atom_pos for p in self.pairs]
-            fit_pos_arrays = [p.fit_molec.atom_pos for p in self.pairs]
-            ref_ch_arrays = [p.ref_molec.partial_charges[p.ref_molec._nonH_atoms_idx]
-                             for p in self.pairs]
-            fit_ch_arrays = [p.fit_molec.partial_charges[p.fit_molec._nonH_atoms_idx]
-                             for p in self.pairs]
-        else:
-            ref_pos_arrays = [p.ref_molec.mol.GetConformer().GetPositions().astype(np.float32)
-                              for p in self.pairs]
-            fit_pos_arrays = [p.fit_molec.mol.GetConformer().GetPositions().astype(np.float32)
-                              for p in self.pairs]
-            ref_ch_arrays = [p.ref_molec.partial_charges for p in self.pairs]
-            fit_ch_arrays = [p.fit_molec.partial_charges for p in self.pairs]
-
-        max_ref_len = max(a.shape[0] for a in ref_pos_arrays)
-        max_fit_len = max(a.shape[0] for a in fit_pos_arrays)
-
-        entries = []
-        for ref_pos, fit_pos, ref_ch, fit_ch in zip(ref_pos_arrays, fit_pos_arrays,
-                                                      ref_ch_arrays, fit_ch_arrays):
-            orig_ref = ref_pos.shape[0]
-            orig_fit = fit_pos.shape[0]
-
-            ref_pos_pad = np.zeros((max_ref_len, 3), dtype=np.float32)
-            ref_pos_pad[:orig_ref] = ref_pos
-            mask_ref = np.zeros(max_ref_len, dtype=np.float32)
-            mask_ref[:orig_ref] = 1.0
-
-            fit_pos_pad = np.zeros((max_fit_len, 3), dtype=np.float32)
-            fit_pos_pad[:orig_fit] = fit_pos
-            mask_fit = np.zeros(max_fit_len, dtype=np.float32)
-            mask_fit[:orig_fit] = 1.0
-
-            ref_ch_pad = np.zeros(max_ref_len, dtype=np.float32)
-            ref_ch_pad[:orig_ref] = ref_ch
-            fit_ch_pad = np.zeros(max_fit_len, dtype=np.float32)
-            fit_ch_pad[:orig_fit] = fit_ch
-
-            entries.append((ref_pos_pad, fit_pos_pad, ref_ch_pad, fit_ch_pad,
-                            mask_ref, mask_fit, orig_ref, orig_fit))
-
-        return entries, max_ref_len, max_fit_len
-
-    def align_with_vol_esp(self,
-                           lam: float,
-                           no_H: bool = True,
-                           num_repeats: int = 50,
-                           trans_init: bool = False,
-                           lr: float = 0.1,
-                           max_num_steps: int = 200,
-                           verbose: bool = False
-                           ) -> Tuple[np.ndarray, List[np.ndarray]]:
-        """Align all pairs using padded masked volumetric ESP similarity via JAX.
-
-        Because all padded arrays have the same shape, JAX's XLA compiler
-        reuses one compiled kernel for every pair — no recompilation overhead.
-
-        Results are stored in-place on each MoleculePair:
-        - ``pair.transform_vol_esp_noH`` / ``pair.sim_aligned_vol_esp_noH`` (when ``no_H=True``)
-        - ``pair.transform_vol_esp``     / ``pair.sim_aligned_vol_esp``      (when ``no_H=False``)
-
-        Parameters
-        ----------
-        lam : float
-            Partial charge weighting parameter. Typically 0.1 for volumetric.
-        no_H : bool
-            Whether to exclude hydrogens. Default is True.
-        num_repeats : int
-            Number of SE(3) initializations per pair. Default is 50.
-        trans_init : bool
-            If True, initialize translations to each ref atom position. Default is False.
-        lr : float
-            Optimizer learning rate. Default is 0.1.
-        max_num_steps : int
-            Maximum optimization steps. Default is 200.
-        verbose : bool
-            Print scores per pair. Default is False.
-
-        Returns
-        -------
-        scores : np.ndarray
-            Scores for each pair. Shape: (N,).
-        aligned_list : list of np.ndarray
-            Aligned fit atom coordinates (unpadded) for each pair.
-        """
-        try:
-            import jax.numpy as jnp
-        except ImportError as exc:
-            raise ImportError(
-                'JAX is required for MoleculePairBatch.align_with_vol_esp. '
-                'Install it with: pip install "shepherd-score[jax]"'
-            ) from exc
-
-        from .alignment_jax import optimize_ROCS_esp_overlay_jax_mask
-
-        entries, _, _ = self._pad_and_mask_vol_esp(no_H=no_H)
-        aligned_list = []
-        scores = np.zeros((len(self.pairs),))
-
-        for i, (pair, entry) in enumerate(zip(self.pairs, entries)):
-            (ref_pos_pad, fit_pos_pad, ref_ch_pad, fit_ch_pad,
-             mask_ref, mask_fit, orig_ref, orig_fit) = entry
-
-            trans_centers = None
-            if trans_init:
-                if no_H:
-                    trans_centers = pair.ref_molec.atom_pos
-                else:
-                    trans_centers = pair.ref_molec.mol.GetConformer().GetPositions().astype(np.float32)
-
-            aligned_pts, se3_transform, score = optimize_ROCS_esp_overlay_jax_mask(
-                ref_points=jnp.array(ref_pos_pad),
-                fit_points=jnp.array(fit_pos_pad),
-                ref_charges=jnp.array(ref_ch_pad),
-                fit_charges=jnp.array(fit_ch_pad),
-                mask_ref=jnp.array(mask_ref),
-                mask_fit=jnp.array(mask_fit),
-                alpha=0.81,
-                lam=lam,
-                num_repeats=num_repeats,
-                trans_centers=trans_centers,
-                lr=lr,
-                max_num_steps=max_num_steps,
-                verbose=verbose,
-            )
-
-            se3_transform = np.array(se3_transform)
-            score = np.array(score)
-            aligned_pts = np.array(aligned_pts)[:orig_fit]
-            scores[i] = score
-
-            if no_H:
-                pair.transform_vol_esp_noH = se3_transform
-                pair.sim_aligned_vol_esp_noH = score
-            else:
-                pair.transform_vol_esp = se3_transform
-                pair.sim_aligned_vol_esp = score
-
-            aligned_list.append(aligned_pts)
-
-        return scores, aligned_list
-
-    def align_with_surf(self,
-                        alpha: float,
-                        num_repeats: int = 50,
-                        trans_init: bool = False,
-                        lr: float = 0.1,
-                        max_num_steps: int = 200,
-                        use_jax: bool = True,
-                        use_analytical: bool = True,
-                        verbose: bool = False
-                        ) -> Tuple[np.ndarray, List[np.ndarray]]:
-        """Align all pairs using surface similarity by delegating to MoleculePair.
-
-        Surface arrays are same-sized across all pairs so no padding is needed.
-        Results are stored in-place on each MoleculePair:
-        - ``pair.transform_surf`` and ``pair.sim_aligned_surf``
-
-        Parameters
-        ----------
-        alpha : float
-            Gaussian width parameter for overlap.
-        num_repeats : int
-            Number of SE(3) initializations per pair. Default is 50.
-        trans_init : bool
-            Apply translation initialization for alignment. Default is False.
-        lr : float
-            Optimizer learning rate. Default is 0.1.
-        max_num_steps : int
-            Maximum optimization steps. Default is 200.
-        use_jax : bool
-            Whether to use JAX backend. Default is True.
-        use_analytical : bool
-            Whether to use analytical gradients (PyTorch only). Default is True.
-        verbose : bool
-            Print scores per pair. Default is False.
-
-        Returns
-        -------
-        scores : np.ndarray
-            Scores for each pair. Shape: (N,).
-        aligned_list : list of np.ndarray
-            Aligned fit surface coordinates for each pair.
-        """
-        aligned_list = []
-        scores = np.zeros((len(self.pairs),))
-
-        for i, pair in enumerate(self.pairs):
-            aligned_pts = pair.align_with_surf(
-                alpha=alpha,
-                num_repeats=num_repeats,
-                trans_init=trans_init,
-                lr=lr,
-                max_num_steps=max_num_steps,
-                use_jax=use_jax,
-                use_analytical=use_analytical,
-                verbose=verbose,
-            )
-            scores[i] = float(pair.sim_aligned_surf)
-            aligned_list.append(aligned_pts)
-
-        return scores, aligned_list
-
-    def align_with_esp(self,
-                       alpha: float,
-                       lam: float = 0.3,
-                       num_repeats: int = 50,
-                       trans_init: bool = False,
-                       lr: float = 0.1,
-                       max_num_steps: int = 200,
-                       use_jax: bool = True,
-                       use_analytical: bool = True,
-                       verbose: bool = False
-                       ) -> Tuple[np.ndarray, List[np.ndarray]]:
-        """Align all pairs using ESP+surface similarity by delegating to MoleculePair.
-
-        Surface arrays are same-sized across all pairs so no padding is needed.
-        Results are stored in-place on each MoleculePair:
-        - ``pair.transform_esp`` and ``pair.sim_aligned_esp``
-
-        Parameters
-        ----------
-        alpha : float
-            Gaussian width parameter for overlap.
-        lam : float
-            Weighting factor for ESP scoring. Scaled internally. Default is 0.3.
-        num_repeats : int
-            Number of SE(3) initializations per pair. Default is 50.
-        trans_init : bool
-            Apply translation initialization for alignment. Default is False.
-        lr : float
-            Optimizer learning rate. Default is 0.1.
-        max_num_steps : int
-            Maximum optimization steps. Default is 200.
-        use_jax : bool
-            Whether to use JAX backend. Default is True.
-        use_analytical : bool
-            Whether to use analytical gradients (PyTorch only). Default is True.
-        verbose : bool
-            Print scores per pair. Default is False.
-
-        Returns
-        -------
-        scores : np.ndarray
-            Scores for each pair. Shape: (N,).
-        aligned_list : list of np.ndarray
-            Aligned fit surface coordinates for each pair.
-        """
-        aligned_list = []
-        scores = np.zeros((len(self.pairs),))
-
-        for i, pair in enumerate(self.pairs):
-            aligned_pts = pair.align_with_esp(
-                alpha=alpha,
-                lam=lam,
-                num_repeats=num_repeats,
-                trans_init=trans_init,
-                lr=lr,
-                max_num_steps=max_num_steps,
-                use_jax=use_jax,
-                use_analytical=use_analytical,
-                verbose=verbose,
-            )
-            scores[i] = float(pair.sim_aligned_esp)
-            aligned_list.append(aligned_pts)
-
-        return scores, aligned_list
-
-    def _pad_and_mask_pharm(self):
-        """Extract, pad, and create masks for pharmacophore alignment.
-
-        Validates that all pairs have pharmacophore data. Does NOT modify the
-        pair objects. Returns padded arrays and masks.
-
-        Returns
-        -------
-        entries : list of tuples
-            Each tuple is (ref_ptypes, fit_ptypes,
-                           ref_ancs_pad, fit_ancs_pad,
-                           ref_vecs_pad, fit_vecs_pad,
-                           mask_ref, mask_fit,
-                           orig_ref_len, orig_fit_len).
-        max_ref_len : int
-        max_fit_len : int
-        """
-        for i, pair in enumerate(self.pairs):
-            if (pair.ref_molec.pharm_types is None or
-                    pair.fit_molec.pharm_types is None):
-                raise ValueError(
-                    f'Pair {i} is missing pharmacophore data. '
-                    'Create Molecule objects with pharm_multi_vector set to True or False.'
-                )
-
-        ref_types_list = [p.ref_molec.pharm_types for p in self.pairs]
-        fit_types_list = [p.fit_molec.pharm_types for p in self.pairs]
-        ref_ancs_list  = [p.ref_molec.pharm_ancs  for p in self.pairs]
-        fit_ancs_list  = [p.fit_molec.pharm_ancs  for p in self.pairs]
-        ref_vecs_list  = [p.ref_molec.pharm_vecs  for p in self.pairs]
-        fit_vecs_list  = [p.fit_molec.pharm_vecs  for p in self.pairs]
-
-        max_ref_len = max(t.shape[0] for t in ref_types_list)
-        max_fit_len = max(t.shape[0] for t in fit_types_list)
-
-        DUMMY_TYPE = 8  # index of 'Dummy' in P_TYPES
-
-        entries = []
-        for (ref_types, fit_types,
-             ref_ancs, fit_ancs,
-             ref_vecs, fit_vecs) in zip(ref_types_list, fit_types_list,
-                                        ref_ancs_list, fit_ancs_list,
-                                        ref_vecs_list, fit_vecs_list):
-            orig_ref = ref_types.shape[0]
-            orig_fit = fit_types.shape[0]
-
-            # types: pad with DUMMY_TYPE
-            ref_types_pad = np.full(max_ref_len, DUMMY_TYPE, dtype=np.int32)
-            ref_types_pad[:orig_ref] = ref_types
-            fit_types_pad = np.full(max_fit_len, DUMMY_TYPE, dtype=np.int32)
-            fit_types_pad[:orig_fit] = fit_types
-
-            # anchors / vectors: pad with zeros
-            ref_ancs_pad = np.zeros((max_ref_len, 3), dtype=np.float32)
-            ref_ancs_pad[:orig_ref] = ref_ancs
-            fit_ancs_pad = np.zeros((max_fit_len, 3), dtype=np.float32)
-            fit_ancs_pad[:orig_fit] = fit_ancs
-
-            ref_vecs_pad = np.zeros((max_ref_len, 3), dtype=np.float32)
-            ref_vecs_pad[:orig_ref] = ref_vecs
-            fit_vecs_pad = np.zeros((max_fit_len, 3), dtype=np.float32)
-            fit_vecs_pad[:orig_fit] = fit_vecs
-
-            # binary masks (float for JAX)
-            mask_ref = np.zeros(max_ref_len, dtype=np.float32)
-            mask_ref[:orig_ref] = 1.0
-            mask_fit = np.zeros(max_fit_len, dtype=np.float32)
-            mask_fit[:orig_fit] = 1.0
-
-            entries.append((ref_types_pad, fit_types_pad,
-                            ref_ancs_pad, fit_ancs_pad,
-                            ref_vecs_pad, fit_vecs_pad,
-                            mask_ref, mask_fit,
-                            orig_ref, orig_fit))
-
-        return entries, max_ref_len, max_fit_len
-
-    def align_with_pharm(self,
-                         similarity: str = 'tanimoto',
-                         extended_points: bool = False,
-                         only_extended: bool = False,
-                         num_repeats: int = 50,
-                         trans_init: bool = False,
-                         lr: float = 0.1,
-                         max_num_steps: int = 200,
-                         verbose: bool = False
-                         ) -> Tuple[np.ndarray, List[np.ndarray], List[np.ndarray]]:
-        """Align all pairs using padded masked pharmacophore similarity via JAX.
-
-        Because all padded arrays have the same shape, JAX's XLA compiler
-        reuses one compiled kernel for every pair — no recompilation overhead.
-
-        Results are stored in-place on each MoleculePair:
-        - ``pair.transform_pharm`` and ``pair.sim_aligned_pharm``
-
-        Parameters
-        ----------
-        similarity : str
-            One of ``'tanimoto'``, ``'tversky'``, ``'tversky_ref'``, ``'tversky_fit'``.
-        extended_points : bool
-            Score HBA/HBD with extended-point Gaussians.
-        only_extended : bool
-            When ``extended_points`` is True, ignore anchor overlaps.
-        num_repeats : int
-            Number of SE(3) initializations per pair.
-        trans_init : bool
-            If True, initialize translations to each ref pharmacophore anchor.
-        lr : float
-            Optimizer learning rate.
-        max_num_steps : int
-            Maximum optimization steps.
-        verbose : bool
-            Print scores per pair.
-
-        Returns
-        -------
-        scores : np.ndarray
-            Shape: (N,).
-        aligned_anchors_list : list of np.ndarray
-            Aligned fit pharmacophore anchors (unpadded) for each pair.
-        aligned_vectors_list : list of np.ndarray
-            Aligned fit pharmacophore vectors (unpadded) for each pair.
-        """
-        try:
-            import jax.numpy as jnp
-        except ImportError as exc:
-            raise ImportError(
-                'JAX is required for MoleculePairBatch.align_with_pharm. '
-                'Install it with: pip install "shepherd-score[jax]"'
-            ) from exc
-
-        from .alignment_jax import optimize_pharm_overlay_jax_vectorized_mask
-
-        entries, _, _ = self._pad_and_mask_pharm()
-        aligned_anchors_list = []
-        aligned_vectors_list = []
-        scores = np.zeros((len(self.pairs),))
-
-        for i, (pair, entry) in enumerate(zip(self.pairs, entries)):
-            (ref_types_pad, fit_types_pad,
-             ref_ancs_pad, fit_ancs_pad,
-             ref_vecs_pad, fit_vecs_pad,
-             mask_ref, mask_fit,
-             orig_ref, orig_fit) = entry
-
-            trans_centers = None
-            if trans_init:
-                trans_centers = pair.ref_molec.pharm_ancs
-
-            aligned_ancs, aligned_vecs, se3_transform, score = optimize_pharm_overlay_jax_vectorized_mask(
-                ref_pharms=jnp.array(ref_types_pad),
-                fit_pharms=jnp.array(fit_types_pad),
-                ref_anchors=jnp.array(ref_ancs_pad),
-                fit_anchors=jnp.array(fit_ancs_pad),
-                ref_vectors=jnp.array(ref_vecs_pad),
-                fit_vectors=jnp.array(fit_vecs_pad),
-                mask_ref=jnp.array(mask_ref),
-                mask_fit=jnp.array(mask_fit),
-                similarity=similarity,
-                extended_points=extended_points,
-                only_extended=only_extended,
-                num_repeats=num_repeats,
-                trans_centers=trans_centers,
-                init_ref_anchors=pair.ref_molec.pharm_ancs,
-                init_fit_anchors=pair.fit_molec.pharm_ancs,
-                lr=lr,
-                max_num_steps=max_num_steps,
-                verbose=verbose,
-            )
-
-            se3_transform = np.array(se3_transform)
-            score = float(np.array(score))
-            aligned_ancs = np.array(aligned_ancs)[:orig_fit]
-            aligned_vecs = np.array(aligned_vecs)[:orig_fit]
-
-            scores[i] = score
-            pair.transform_pharm = se3_transform
-            pair.sim_aligned_pharm = score
-
-            aligned_anchors_list.append(aligned_ancs)
-            aligned_vectors_list.append(aligned_vecs)
-
-        return scores, aligned_anchors_list, aligned_vectors_list
